Crate apic

Structs

ApicRegisters

A structure that offers access to APIC/xAPIC through its I/O registers.

LocalApic

This structure represents a single Local APIC in the system; there is one per CPU.

RegisterArray

Enums

InterruptChip

LapicInitError

The possible errors that can occur in LocalApic::init().

LapicIpiDestination

The possible destination shorthand values for IPI ICR.

LvtLint

The Local APIC’s vector table local interrupt pins.

Constants

APIC_SPURIOUS_INTERRUPT_IRQ

The IRQ number reserved for spurious APIC interrupts (as recommended by OS dev wiki).

LOCAL_APIC_LVT_IRQ

The IRQ number reserved for Local APIC timer interrupts in the IDT.

Statics

INTERRUPT_CHIP

The interrupt chip that is currently configured on this machine. The default is InterruptChip::PIC, but the typical case is APIC or X2APIC, which will be set once those chips have been initialized.

Functions

cpu_count

Returns the number of CPUs (cores, or local APICs) that exist and are currently initialized on this system.

get_bsp_id

get_lapics

Returns a reference to the list of LocalApics, one per CPU core.

get_my_apic

Returns a reference to the LocalApic for the currently executing CPU core.

get_my_apic_id

Returns the APIC ID of the currently executing CPU core.

has_x2apic

Returns true if the machine has support for x2apic

init

Determines whether this system contains an xapic or x2apic and enables the Local APIC hardware in the correct mode.

is_bsp

Returns true if the currently executing processor core is the bootstrap processor, i.e., the first procesor to run