Struct apic::LocalApic

pub struct LocalApic {
    pub regs: Option<BoxRefMut<MappedPages, ApicRegisters>>,
    pub processor: u8,
    pub apic_id: u8,
    pub is_bsp: bool,
}

This structure represents a single APIC in the system, there is one per core.

Fields
regs: Option<BoxRefMut<MappedPages, ApicRegisters>>

Only exists for xapic, should be None for x2apic systems.

processor: u8

The processor id of this APIC.

apic_id: u8

The APIC system id of this APIC.

is_bsp: bool

Whether this LocalApic is the bootstrap processor (the first processor to boot up).

Implementations

impl LocalApic

pub fn new(
page_table: &mut PageTable,
processor: u8,
apic_id: u8,
is_bsp: bool,
nmi_lint: u8,
nmi_flags: u16
) -> Result<LocalApic, &'static str>

This MUST be invoked from the AP core itself when it is booting up. The BSP cannot invoke this for other APs (it can only invoke it for itself).

pub fn id(&self) -> u8

pub fn version(&self) -> u32

pub fn error(&self) -> u32

pub fn clear_error(&mut self)

pub fn icr(&self) -> u64

pub fn set_icr(&mut self, value: u64)

pub fn send_ipi(&mut self, irq: u8, destination: LapicIpiDestination)

Send an IPI to the cores specified by the given destination

pub fn send_nmi_ipi(&mut self, destination: LapicIpiDestination)

Send a NMI IPI to the cores specified by the given destination

pub fn eoi(&mut self)

pub fn set_ldr(&mut self, value: u32)

pub fn set_nmi(&mut self, lint: u8, flags: u16) -> Result<(), &'static str>

Set the NonMaskableInterrupt redirect for this LocalApic. Argument lint can be either 0 or 1, since each local APIC has two LVT LINTs (Local Vector Table Local INTerrupts)

pub fn get_isr(&self) -> [u32; 8]

Returns the values of the 8 in-service registers for this APIC, which is a series of bitmasks that shows which interrupt lines are currently being serviced.

pub fn get_irr(&self) -> [u32; 8]

Returns the values of the 8 request registers for this APIC, which is a series of bitmasks that shows which interrupt lines are currently raised, but not yet being serviced.

pub fn clear_pmi_mask(&mut self)

Clears the interrupt mask bit in the apic performance monitor register.