Struct apic::LocalApic

pub struct LocalApic { /* private fields */ }

This structure represents a single APIC in the system, there is one per core.

Implementations

impl LocalApic

pub fn init(
page_table: &mut PageTable,
processor: u8,
apic_id: u8,
is_bsp: bool,
nmi_lint: u8,
nmi_flags: u16
) -> Result<(), &'static str>

Creates and initializes a new LocalApic for the current CPU core and adds it to the global set of initialized local APICs.

Important Usage Note

This MUST be invoked from the AP core itself when it is booting up. The BSP cannot invoke this for other APs (it can only invoke it for itself).

pub fn apic_id(&self) -> u8

Returns the APIC ID of this lapic.

This value comes from the MADT ACPI table entry that was used to boot up this CPU core.

Currently this returns the same value as LocalApic::id().

pub fn processor(&self) -> u8

Returns the “processor ID” of this lapic, which is usually a meaningless value.

This value comes from the MADT ACPI table entry that was used to boot up this CPU core.

pub fn is_bsp(&self) -> bool

Returns true if this CPU core was the BootStrap Processor (BSP), i.e., the first CPU to boot and run the OS code.

There is only one BSP per system.

pub fn enable_lvt_timer(&mut self, enable: bool)

Enable (unmask) or disable (mask) the LVT timer interrupt on this lapic.

This does not modify the timer’s current count value.

pub fn id(&self) -> u8

Reads the hardware-provided ID of this lapic (from its registers).

Currently, this matches the value returned by LocalApic::apic_id().

pub fn version(&self) -> u32

Returns the version of this lapic.

pub fn error(&self) -> u32

Returns the value of this lapic’s error register.

pub fn clear_error(&mut self)

Clears/resets this lapic’s error register.

pub fn icr(&self) -> u64

Reads the current value of this lapic’s Interrupt Control Register.

pub fn set_icr(&mut self, value: u64)

Writes value to this lapic’s Interrupt Control Register.

pub fn send_ipi(&mut self, irq: u8, destination: LapicIpiDestination)

Send an IPI to the cores specified by the given destination

pub fn send_nmi_ipi(&mut self, destination: LapicIpiDestination)

Send a NMI IPI to the cores specified by the given destination

pub fn eoi(&mut self)

Send an End Of Interrupt (EOI) signal to this local APIC, which indicates that the calling interrupt handler has finished handling the current interrupt.

pub fn set_nmi(&mut self, lint: LvtLint, flags: u16)

Set the NonMaskableInterrupt redirect for this LocalApic. Argument lint can be either 0 or 1, since each local APIC has two LVT LINTs (Local Vector Table Local INTerrupts)

pub fn get_isr(&self) -> [u32; 8]

Returns the values of the 8 in-service registers for this APIC, which is a series of bitmasks that shows which interrupt lines are currently being serviced.

pub fn get_irr(&self) -> [u32; 8]

Returns the values of the 8 request registers for this APIC, which is a series of bitmasks that shows which interrupt lines are currently raised, but not yet being serviced.

pub fn clear_pmi_mask(&mut self)

Clears the interrupt mask bit in the apic performance monitor register.